// front end/src/app/model/customerdetails.model.ts
export class Customerdetails {
     cust_local_ps: string ;
     cust_localdistid: string ;
       cust_localstateid: string;
   cust_p_address: string;
   cust_p_ps: string;
 cust_p_distid: string;
 cust_p_stateid: string;
 annual_income: string;
   cust_p_d_id: number;
    cust_id: number;
     cust_occupation: string;
 cust_local_address: string; }